[#IOPID-857] minor refactor
