+++ src/functions/__mocks__/general.ts
@@ -1,0 +1,4 @@
+import { FiscalCode } from "@pagopa/ts-commons/lib/strings";
+
+export const aFiscalCode = "AAAAAA89S20I111X" as FiscalCode;
+export const anotherFiscalCode = "AAAAAA89S20I111Z" as FiscalCode;
